% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Squaring Numeric Data}
\usage{
cube(x)
}
\arguments{
\item{x}{numeric, vector, number(s) to be cubed}
}
\value{
numeric. vector
}
\description{
This function will return the input cubed. Can be numeric or vector.
}
\examples{
cube(5)
cube(0)
cube(c(1,2,3))
cube(d)

}
