% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squaring Numeric Data}
\usage{
square(x)
}
\arguments{
\item{x}{numeric, vector, number(s) to be squared}
}
\value{
numeric. vector
}
\description{
This function will return the input squared. Can be numeric or vector.
}
\examples{
square(5)
square(0)
square(c(1,2,3))
square(d)

}
